Created modular CardList and wrapper components
Upcoming Deadlines (todos) and daily schedule are now wrapper classes
implementing the general CardList component
