+++ app/App.tsx
@@ -11,7 +11,6 @@ import { EvaIconsPack } from '@ui-kitten/eva-icons';
 import rootReducer from './redux/reducers/RootReducer';
 
 import AuthContainer from './AuthContainer';
-import HomePage from './screens/main/HomePage';
 
 type Props = {};
 
@@ -22,8 +21,7 @@ const App: React.FC<Props> = () => {
     <Provider store={store}>
       <IconRegistry icons={EvaIconsPack} />
       <ApplicationProvider {...eva} theme={eva.light}>
-        {/* <AuthContainer /> */}
-        <HomePage />
+        <AuthContainer />
       </ApplicationProvider>
     </Provider>
   );
