+++ app/components/todos/DailySchedule.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Text, ListItem, List, Divider } from '@ui-kitten/components';
+import CardList from "./CardList";
 
 type scheduleItem = {
   title: String;
@@ -10,10 +11,11 @@ type scheduleItem = {
   date: Date;
 };
 
-type renderitem = {
-  item: scheduleItem;
-  index: Number;
-};
+type displayableItem = {
+  name: String;
+  prefix: String;
+  desc: String;
+}
 
 type propObject = {
   schedule: Array<scheduleItem>;
@@ -21,44 +23,27 @@ type propObject = {
 };
 
 const DailySchedule = ({ schedule, handleSchedule }: propObject) => {
-  let displaySchedule;
 
-  const filterByDate = (item: scheduleItem) : boolean => {
+  const filterSched = (item: scheduleItem) : boolean => {
     const today = new Date();
     const itemDate = item.date;
     return itemDate.getDate() === today.getDate() && itemDate.getMonth() === today.getMonth() && itemDate.getFullYear() === today.getFullYear();
   };
 
-  const renderItem = ({ item, index }: renderitem) => {
-    return (
-      <ListItem
-        title={`${item.startHour}:${
-          item.startMinute === 0 ? '00' : item.startMinute
-        } - ${item.endHour}:${item.endMinute === 0 ? '00' : item.endMinute} : ${
-          item.title
-        }`}
-        // onPress={() => handleTodo(todos.filter((td) => td.id !== item.id))} // can add deletion or editing features in the future
-      />
-    );
+  const formatSched = (item: scheduleItem) : displayableItem => {
+    return {
+      prefix: `${item.startHour}:${item.startMinute === 0 ? '00' : item.startMinute
+        } - ${item.endHour}:${item.endMinute === 0 ? '00' : item.endMinute}`,
+      name: item.title,
+      desc: ''
+    };
   };
 
-  if (schedule === undefined || schedule === []) {
-    displaySchedule = (
-      <Text appearance="hint" category="c2">
-        You have no to dos! Click the button above to start adding some!
-      </Text>
-    );
-  } else {
-    displaySchedule = (
-      <List
-        data={schedule.filter(item => filterByDate(item))}
-        ItemSeparatorComponent={Divider}
-        renderItem={renderItem}
-      />
-    );
-  }
+  const formatFilterSched = () => {
+    return schedule.filter(filterSched).map(formatSched);
+  };
 
-  return displaySchedule;
+  return <CardList data={formatFilterSched()} />;
 };
 
 export default DailySchedule;
