+++ app/components/todos/UpcomingDeadlines.tsx
@@ -1,54 +1,39 @@
 import React from 'react';
 import { ListItem, List, Text, Divider } from '@ui-kitten/components';
+import CardList from "./CardList";
 
-type todoObject = {
+type PropObject = {
+  todos: Array<todoItem>;
+  handleTodo: Function;
+};
+
+type displayableItem = {
+  name: String;
+  prefix: String;
+  desc: String;
+}
+
+type todoItem = {
   id: number;
   title: string;
   desc: string;
   date: Date;
 };
 
-type renderitem = {
-  item: todoObject;
-  index: number;
-};
-
-type PropObject = {
-  todos: Array<todoObject>;
-  handleTodo: Function;
-};
-
 const UpcomingDeadlines = ({ todos, handleTodo }: PropObject) => {
-  let todoList;
-  const renderItem = ({ item, index }: renderitem) => {
-    return (
-      <ListItem
-        title={`${item.date.getDate()}/${
+
+  const formatTodo = (item: todoItem) : displayableItem => {
+    return {
+      prefix: `${item.date.getDate()}/${
           item.date.getMonth() + 1
-        }/${item.date.getFullYear()}: ${item.title}`}
-        description={item.desc}
-        onPress={() => handleTodo(todos.filter((td) => td.id !== item.id))}
-      />
-    );
+        }/${item.date.getFullYear()}`,
+      name: item.title,
+      desc: item.desc
+    };
   };
 
-  if (todos === undefined || todos === []) {
-    todoList = (
-      <Text appearance="hint" category="c2">
-        You have no to dos! Click the button above to start adding some!
-      </Text>
-    );
-  } else {
-    todoList = (
-      <List
-        data={todos}
-        ItemSeparatorComponent={Divider}
-        renderItem={renderItem}
-      />
-    );
-  }
 
-  return todoList;
+  return <CardList data={todos.map(formatTodo)}/>;
 };
 
 export default UpcomingDeadlines;
